pyproject.toml
[project]
name = "hua-news-ai-rag"
version = "0.1.0"
description = "just use to test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "flask>=3.0.0",
  "flask-cors>=4.0.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.7.0",
  "pyjwt>=2.8.0",
  "passlib[bcrypt]>=1.7.4",
  "feedparser>=6.0.11",
  "requests>=2.32.0",
  "numpy==1.26.4",
  "sentence-transformers==2.7.0",
  "faiss-cpu==1.8.0",
  "apscheduler>=3.11.0",
]

[project.optional-dependencies]
embeddings = [
  "sentence-transformers>=2.7.0",
  "faiss-cpu>=1.8.0",
  "numpy>=1.26.0"
]
langchain = [
  "langchain>=0.2.12",
  "langchain-community>=0.2.10",
  "ollama>=0.3.0",
  "tiktoken>=0.7.0"
]
